requirements.txt
Flask
spotipy
python-dotenv
pytest
gunicorn
Flask-Limiter
redis
Flask-Caching